requirements.txt
astroid==2.4.2
audioread==2.1.8
cffi==1.14.0
decorator==4.4.2
ffmpeg-python==0.2.0
future==0.18.2
isort==4.3.21
joblib==0.15.1
lazy-object-proxy==1.4.3
librosa==0.7.2
llvmlite==0.31.0
mccabe==0.6.1
numba==0.48.0
numpy==1.18.5
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.3
pytube3==9.6.4
resampy==0.2.2
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
wrapt==1.12.1
